requirements.txt
paho-mqtt
numpy
tqdm
requests